What is overkill?
overkill. v., n. -v. 1. to obliterate (a target) with more nuclear force
than required. -n. 1. a destructive capacity greatly exceeding that
required for a given target. 2. an excess of something (such as a
quantity or an action) beyond what is required or suitable for a
particular purpose. 3. killing in excess of what is intended or
required.